Move `throttled_loop` to util module
Same deal as previous commit. Avoid single-function modules.
